Renamed the FormData class to DiscoverFiltersFormData to be more specific of it's purpose
